+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+# Dev Update v1.1.1
+- Added logic for determining if we're using the Production bot or the Development bot
+- Added more console logging
+- Fixed small lang issue with the `/get-roles` command
+
 # Dev Update v1.1.0
 - First 'Beta' version! (I'm putting this on a larger server)
+- I also made a dev bot so I do mess anything with the main bot.
 - Fixed tickets so anyone that can view the ticket can close/admin only it (since the logic to make sure only specific roles see tickets is in place)
 
 # Dev Update v1.0.3